requirements.txt
beautifulsoup4==4.13.3
chromadb==0.6.3
gradio==5.22.0
huggingface_hub==0.29.3
llama_index==0.12.25
nest_asyncio==1.6.0
openai==1.68.0
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
tiktoken==0.9.0
tqdm==4.67.1
